api/requirements.txt
flask
google-cloud-bigquery